Part B/requirements.txt
selenium==4.27.1
pytest==8.3.4
pytest-html==4.1.1
webdriver-manager==4.0.2
requests==2.32.3
python-dotenv==1.0.1
